\name{Ibasam-package}
\alias{Ibasam-package}
\alias{Ibasam}
\docType{package}
\title{
Ibasam2
}
\description{
Ibasam2
}
\details{
\tabular{ll}{
Package: \tab Ibasam\cr
Type: \tab Package\cr
Version: \tab 2.3.0\cr
Date: \tab 2015-09-05\cr
License: \tab GPL2\cr
}
}
\author{
Cyril Piou <piou.cyril@yahoo.fr>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
